+++ types/index.d.ts
@@ -44,4 +44,5 @@ export {default as union} from './union';
 export {default as update} from './update';
 export {default as validate} from './validate';
 export {default as zip} from './zip';
+export {default as exists} from './exists';
 export * from './types';